requirements.txt
pytorch==1.2.0
torchvision==0.4.0
cudatoolkit==10.0
tensorboard==1.14
tensorflow==1.14
torchtext==0.4.0
nltk
spacy
transformers
datasets
sentencepiece